plugins: separate the portaudio into sub-modules
The current portaudio plugin includes both recorder and pcm player
drivers. It even includes both aync and sync methods.

Since maintenance is difficult in this state, the portaudio plugin
is separated as follows:

* `portaudio.so`

  Perform portaudio Pa_Initialize() / Pa_Terminate().
  The plugin has high priority to initialize first.

* `portaudio_recorder.so`

  Recorder driver function only.

* `portaudio_pcm_async.so`:

  Performs the pcm play driver function.
  Default pcm driver.

* `portaudio_pcm_sync.so`:

  performs the pcm play driver function.
  Sync method for low performance devices.

For testing, provide a following environment variable to override
the default pcm driver.

    export NUGU_DEFAULT_PCM_DRIVER=portaudio_pcm_sync
